Restore same module structure -- it was needed to avoid cyclical imports
