+++ packages/liveblocks-devtools/src/background.ts
@@ -1,4 +1,4 @@
-import type { DevTools } from "@liveblocks/core";
+import type { DevToolsMsg } from "@liveblocks/core";
 import type { Runtime } from "webextension-polyfill";
 import browser from "webextension-polyfill";
 
@@ -9,7 +9,7 @@ const portsByTabId: Map<number, Runtime.Port> = new Map();
  * running the client.
  */
 browser.runtime.onConnect.addListener((port) => {
-  function handleMessage(message: DevTools.FullPanelToClientMessage) {
+  function handleMessage(message: DevToolsMsg.FullPanelToClientMessage) {
     //
     // NOTE: Special eaves dropping happening here when the panel sends their
     // "connect" message. While this message is intended to signify to the
@@ -44,7 +44,7 @@ browser.runtime.onConnect.addListener((port) => {
  * routes the message to the correct dev panel window.
  */
 browser.runtime.onMessage.addListener(
-  (message: DevTools.FullClientToPanelMessage, sender) => {
+  (message: DevToolsMsg.FullClientToPanelMessage, sender) => {
     const tabId = sender.tab?.id;
     if (tabId) {
       portsByTabId.get(tabId)?.postMessage(message);
